backend/requirements.txt
fastapi
uvicorn[standard]
mediapipe
opencv-python
numpy
pillow
pydantic>=2.0
pydantic-settings>=2.0
httpx>=0.27
fal-client>=0.4
openai>=1.13
